+++ components/info/impl/src/main/java/com/flipperdevices/info/impl/di/InfoComponent.kt
@@ -7,6 +7,7 @@ import com.flipperdevices.info.impl.viewmodel.ConnectViewModel
 import com.flipperdevices.info.impl.viewmodel.DeviceInfoViewModel
 import com.flipperdevices.info.impl.viewmodel.DeviceStatusViewModel
 import com.flipperdevices.info.impl.viewmodel.FirmwareUpdateViewModel
+import com.flipperdevices.info.impl.viewmodel.FlipperColorViewModel
 import com.squareup.anvil.annotations.ContributesTo
 
 @ContributesTo(AppGraph::class)
@@ -16,5 +17,6 @@ interface InfoComponent {
     fun inject(viewModel: ConnectViewModel)
     fun inject(viewModel: AlarmViewModel)
     fun inject(viewModel: DeviceInfoViewModel)
+    fun inject(viewModel: FlipperColorViewModel)
     fun inject(fragment: InfoFragment)
 }