+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [Feature] Add black flipper support (update and info screen)
 - [Feature] Compose Theme(module `core:ui:theme`, module `archive`)
+- [Feature] Auto disconnect after 5 minutes inactivity
 - [BUGFIX] Lock portrait orientation
 - [BUGFIX] Update card not shown when flipper not connected
 - [BUGFIX] Static URL on dynamic link docs